Refactoring and moving to components
