+++ public/app/features/dashboard/components/PanelEditor/OptionsPaneContent.tsx
@@ -15,35 +15,15 @@ import {
   Input,
 } from '@grafana/ui';
 import { DefaultFieldConfigEditor, OverrideFieldConfigEditor } from './FieldConfigEditor';
-import { AngularPanelOptions } from './AngularPanelOptions';
 import { css } from 'emotion';
-import { GeneralPanelOptions } from './GeneralPanelOptions';
-import { PanelOptionsEditor } from './PanelOptionsEditor';
+import { PanelOptionsTab } from './PanelOptionsTab';
 import { DashNavButton } from 'app/features/dashboard/components/DashNav/DashNavButton';
 
-const tabSelections: Array<SelectableValue<string>> = [
-  {
-    label: 'Options',
-    value: 'options',
-  },
-  {
-    label: 'Fields',
-    value: 'defaults',
-  },
-  {
-    label: 'Overrides',
-    value: 'overrides',
-  },
-  {
-    label: 'General',
-    value: 'panel',
-  },
-];
-
 export const OptionsPaneContent: React.FC<{
-  plugin?: PanelPlugin;
+  plugin: PanelPlugin;
   panel: PanelModel;
   data: PanelData;
+  width: number;
   dashboard: DashboardModel;
   onClose: () => void;
   onFieldConfigsChange: (config: FieldConfigSource) => void;
@@ -53,6 +33,7 @@ export const OptionsPaneContent: React.FC<{
   plugin,
   panel,
   data,
+  width,
   onFieldConfigsChange,
   onPanelOptionsChanged,
   onPanelConfigChange,
@@ -63,17 +44,6 @@ export const OptionsPaneContent: React.FC<{
   const styles = getStyles(theme);
   const [activeTab, setActiveTab] = useState('options');
   const [isSearching, setSearchMode] = useState(false);
-
-  const renderPanelOptions = useCallback(
-    (plugin: PanelPlugin) => {
-      return (
-        <Container padding="md">
-          <>{renderCustomPanelSettings(plugin)}</>
-        </Container>
-      );
-    },
-    [data, plugin, panel, onFieldConfigsChange]
-  );
 
   const renderFieldOptions = useCallback(
     (plugin: PanelPlugin) => {
@@ -118,49 +88,54 @@ export const OptionsPaneContent: React.FC<{
     [data, plugin, panel, onFieldConfigsChange]
   );
 
-  const renderCustomPanelSettings = useCallback(
-    (plugin: PanelPlugin) => {
-      const editors: JSX.Element[] = [];
-      if (plugin.editor && panel) {
-        editors.push(
-          <div className={styles.legacyOptions} key="plugin custom panel settings">
-            <plugin.editor
-              data={data}
-              options={panel.getOptions()}
-              onOptionsChange={onPanelOptionsChanged}
-              fieldConfig={panel.getFieldConfig()}
-              onFieldConfigChange={onFieldConfigsChange}
+  return (
+    <div className={styles.panelOptionsPane}>
+      {plugin && (
+        <div className={styles.wrapper}>
+          <TabsBar className={styles.tabsBar}>
+            <TabsBarContent
+              width={width}
+              isSearching={isSearching}
+              styles={styles}
+              activeTab={activeTab}
+              onClose={onClose}
+              setSearchMode={setSearchMode}
+              setActiveTab={setActiveTab}
             />
-          </div>
-        );
-      }
-
-      // When editor created declaratively
-      if (plugin.optionEditors && panel) {
-        editors.push(
-          <PanelOptionsEditor
-            key="panel options"
-            options={panel.getOptions()}
-            onChange={onPanelOptionsChanged}
-            plugin={plugin}
-          />
-        );
-      }
-
-      if (editors.length > 0) {
-        return editors;
-      }
-
-      return (
-        <div className={styles.legacyOptions}>
-          <AngularPanelOptions panel={panel} dashboard={dashboard} plugin={plugin} />
+          </TabsBar>
+          <TabContent className={styles.tabContent}>
+            <CustomScrollbar>
+              {activeTab === 'options' && (
+                <PanelOptionsTab
+                  panel={panel}
+                  plugin={plugin}
+                  dashboard={dashboard}
+                  data={data}
+                  onPanelConfigChange={onPanelConfigChange}
+                  onFieldConfigsChange={onFieldConfigsChange}
+                  onPanelOptionsChanged={onPanelOptionsChanged}
+                />
+              )}
+              {activeTab === 'defaults' && renderFieldOptions(plugin)}
+              {activeTab === 'overrides' && renderFieldOverrideOptions(plugin)}
+            </CustomScrollbar>
+          </TabContent>
         </div>
-      );
-    },
-    [data, plugin, panel, onFieldConfigsChange]
+      )}
+    </div>
   );
+};
 
-  const renderSearchInput = useCallback(() => {
+export const TabsBarContent: React.FC<{
+  width: number;
+  isSearching: boolean;
+  activeTab: string;
+  styles: OptionsPaneStyles;
+  onClose: () => void;
+  setSearchMode: (mode: boolean) => void;
+  setActiveTab: (tab: string) => void;
+}> = ({ width, isSearching, activeTab, onClose, setSearchMode, setActiveTab, styles }) => {
+  if (isSearching) {
     const defaultStyles = {
       transition: 'width 50ms ease-in-out',
       width: '50%',
@@ -193,74 +168,65 @@ export const OptionsPaneContent: React.FC<{
         }}
       </Transition>
     );
-  }, []);
-
-  const isNarrow = false;
+  }
 
   return (
-    <div className={styles.panelOptionsPane}>
-      {plugin && (
-        <div className={styles.wrapper}>
-          <TabsBar className={styles.tabsBar}>
-            {isSearching && renderSearchInput()}
-            {!isSearching && (
-              <>
-                {isNarrow ? (
-                  <div className="flex-grow-1">
-                    <Select
-                      options={tabSelections}
-                      value={tabSelections.find(v => v.value === activeTab)}
-                      onChange={v => {
-                        setActiveTab(v.value);
-                      }}
-                    />
-                  </div>
-                ) : (
-                  <>
-                    {tabSelections.map(item => {
-                      return (
-                        <Tab
-                          label={item.label}
-                          active={activeTab === item.value}
-                          onChangeTab={() => setActiveTab(item.value)}
-                        />
-                      );
-                    })}
-                    <div className="flex-grow-1" />
-                  </>
-                )}
-                <div className={styles.tabsButton}>
-                  <DashNavButton
-                    icon="fa fa-search"
-                    tooltip="Search all options"
-                    classSuffix="search-options"
-                    onClick={() => setSearchMode(true)}
-                  />
-                </div>
-                <div className={styles.tabsButton}>
-                  <DashNavButton
-                    icon="fa fa-chevron-right"
-                    tooltip="Close options pane"
-                    classSuffix="close-options"
-                    onClick={onClose}
-                  />
-                </div>
-              </>
-            )}
-          </TabsBar>
-          <TabContent className={styles.tabContent}>
-            <CustomScrollbar>
-              {activeTab === 'options' && renderPanelOptions(plugin)}
-              {activeTab === 'defaults' && renderFieldOptions(plugin)}
-              {activeTab === 'overrides' && renderFieldOverrideOptions(plugin)}
-              {activeTab === 'panel' && <GeneralPanelOptions panel={panel} onPanelConfigChange={onPanelConfigChange} />}
-            </CustomScrollbar>
-          </TabContent>
+    <>
+      {width < 377 ? (
+        <div className="flex-grow-1">
+          <Select
+            options={tabSelections}
+            value={tabSelections.find(v => v.value === activeTab)}
+            onChange={v => {
+              setActiveTab(v.value);
+            }}
+          />
         </div>
+      ) : (
+        <>
+          {tabSelections.map(item => {
+            return (
+              <Tab label={item.label} active={activeTab === item.value} onChangeTab={() => setActiveTab(item.value)} />
+            );
+          })}
+          <div className="flex-grow-1" />
+        </>
       )}
-    </div>
+
+      <div className={styles.tabsButton}>
+        <DashNavButton
+          icon="fa fa-search"
+          tooltip="Search all options"
+          classSuffix="search-options"
+          onClick={() => setSearchMode(true)}
+        />
+      </div>
+      <div className={styles.tabsButton}>
+        <DashNavButton
+          icon="fa fa-chevron-right"
+          tooltip="Close options pane"
+          classSuffix="close-options"
+          onClick={onClose}
+        />
+      </div>
+    </>
   );
 };
+
+const tabSelections: Array<SelectableValue<string>> = [
+  {
+    label: 'Options',
+    value: 'options',
+  },
+  {
+    label: 'Fields',
+    value: 'defaults',
+  },
+  {
+    label: 'Overrides',
+    value: 'overrides',
+  },
+];
 
 const getStyles = stylesFactory((theme: GrafanaTheme) => {
   return {
@@ -323,3 +289,5 @@ const getStyles = stylesFactory((theme: GrafanaTheme) => {
     `,
   };
 });
+
+type OptionsPaneStyles = ReturnType<typeof getStyles>;
