+++ public/app/features/dashboard/components/PanelEditor/PanelEditor.tsx
@@ -30,11 +30,6 @@ import { VariableModel } from 'app/features/templating/types';
 import { getVariables } from 'app/features/variables/state/selectors';
 import { SubMenuItems } from 'app/features/dashboard/components/SubMenu/SubMenuItems';
 
-enum Pane {
-  Right,
-  Top,
-}
-
 interface OwnProps {
   dashboard: DashboardModel;
   sourcePanel: PanelModel;
@@ -255,7 +250,7 @@ export class PanelEditorUnconnected extends PureComponent<Props> {
   }
 
   renderOptionsPane() {
-    const { plugin, dashboard, data, panel } = this.props;
+    const { plugin, dashboard, data, panel, uiState } = this.props;
 
     if (!plugin) {
       return <div />;
@@ -267,6 +262,7 @@ export class PanelEditorUnconnected extends PureComponent<Props> {
         dashboard={dashboard}
         data={data}
         panel={panel}
+        width={uiState.rightPaneSize as number}
         onClose={this.onTogglePanelOptions}
         onFieldConfigsChange={this.onFieldConfigChange}
         onPanelOptionsChanged={this.onPanelOptionsChanged}
@@ -341,6 +337,11 @@ const mapDispatchToProps: MapDispatchToProps<DispatchProps, OwnProps> = {
 };
 
 export const PanelEditor = connect(mapStateToProps, mapDispatchToProps)(PanelEditorUnconnected);
+
+enum Pane {
+  Right,
+  Top,
+}
 
 /*
  * Styles
